- setup proguard; - metric module refactor
