+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/telemetry/events/NavigationArriveEvent.kt
@@ -1,10 +1,8 @@
 package com.mapbox.navigation.core.telemetry.events
 
 import android.annotation.SuppressLint
-import androidx.annotation.Keep
 import com.mapbox.navigation.base.metrics.NavigationMetrics
 
-@Keep
 @SuppressLint("ParcelCreator")
 internal class NavigationArriveEvent(
     phoneState: PhoneState
