+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/telemetry/events/RerouteEvent.kt
@@ -1,9 +1,7 @@
 package com.mapbox.navigation.core.telemetry.events
 
-import androidx.annotation.Keep
 import com.mapbox.android.telemetry.TelemetryUtils.obtainUniversalUniqueIdentifier
 
-@Keep
 class RerouteEvent(
     override var sessionState: SessionState
 ) : TelemetryEvent {
