+++ libnavigation-ui/proguard-rules.pro
@@ -1,0 +1,2 @@
+# --- Picasso ---
+-dontwarn com.squareup.okhttp.**
