+++ libnavigator/proguard-rules.pro
@@ -1,0 +1,3 @@
+# --- Navigator ---
+-keep class com.mapbox.navigator.** { *; }
+-keep class com.mapbox.bindgen.** { *; }
