+++ proguard/proguard-project.pro
@@ -1,0 +1,5 @@
+-verbose
+
+# We should at least keep the Exceptions, InnerClasses,
+# and Signature attributes when processing a library.
+-keepattributes Exceptions,InnerClasses,Signature,RuntimeVisibleParameterAnnotations,RuntimeVisibleAnnotations
